Moved TenderFilter to filters.py
Updated mentions of TenderFilter in other areas so this doesn't break anything
